mcp_server/requirements.txt
mcp>=1.0.0
pyyaml>=6.0.0
pydantic>=2.0.0
